requirements.txt
fastapi
uvicorn
psycopg2-binary
sqlalchemy
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-multipart
python-jose[cryptography]
 pydantic[email]
